Move users PDF export from Add User page to Settings admin tab.
Extract shared usersReportPdf helper and wire download through ProfileSettings instead of AddUserPage.
